v0.1.10 Parametrize inputs

+++ backend/src/main/java/fyi/lnz/psych_constructs/database/InsertResult.java
@@ -2,5 +2,5 @@ package fyi.lnz.psych_constructs.database;
 
 import java.util.List;
 
-public record InsertResult(Integer rows, List<Long> generated_keys) {
+public record InsertResult(Integer rows, List<Integer> generated_keys) {
 }
